# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TatkalDoctor Doctor Listing API",
        "description": "Core registry API for the TatkalDoctor ecosystem.\n\n## Authentication\n\n### Public Endpoints\n`GET /api/v1/health` is public — no authentication required.\n\n### Admin Login (Sanctum)\nCall `POST /api/v1/login` with email + password to receive a Bearer token.\n\n### HMAC-SHA256 (All Data Endpoints)\nAll master data and listing endpoints require four request headers:\n\n| Header | Description |\n|---|---|\n| `X-Api-Key` | Client API key issued by super admin |\n| `X-Timestamp` | Unix timestamp in seconds (UTC) |\n| `X-Nonce` | Random string ≥ 8 chars, unique per request |\n| `X-Signature` | HMAC-SHA256 hex-encoded signature |\n\n**String to sign** (fields joined by `\\n`):\n\n```\nHTTP_METHOD\nTIMESTAMP\nNONCE\nREQUEST_PATH\nSHA256(body)\n```\n\n> Timestamp must be within **±5 minutes** of server time. The `REQUEST_PATH` is the URL path only — no query string. For GET requests, `SHA256(body)` is the SHA-256 of an empty string.",
        "contact": {
            "name": "TatkalDoctor Admin",
            "email": "superadmin@tatkaldoctor.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server (WAMP)"
        }
    ],
    "paths": {
...
        "/api/v1/countries": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "List all countries",
                "description": "Returns all countries available on the platform, sorted A→Z. Requires HMAC authentication.",
                "operationId": "getCountries",
                "responses": {
                    "200": {
                        "description": "Countries fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Countries fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Country"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/cities/{country_code}": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "List active cities by country",
                "description": "Returns all active cities for the given ISO alpha-3 country code, sorted A→Z. Requires HMAC authentication.",
                "operationId": "getCitiesByCountry",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "path",
                        "description": "ISO 3166-1 alpha-3 country code (case-insensitive, e.g. IND, USA, GBR)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "IND",
                            "maxLength": 3,
                            "minLength": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cities fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cities fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/City"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Country code not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Country code \"XYZ\" not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/locations/{city_id}": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "List active locations by city",
                "description": "Returns all active locations for the given city ID, sorted A→Z. Requires HMAC authentication.",
                "operationId": "getLocationsByCity",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "path",
                        "description": "ID of the city (from /cities/{country_code})",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locations fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Locations fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/services": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "List active services with sub-services",
                "description": "Returns all active parent services with their active child (sub) services, sorted A→Z. Use service IDs to filter doctor listings. Requires HMAC authentication.",
                "operationId": "getServices",
                "responses": {
                    "200": {
                        "description": "Services fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Services fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Service"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/qualifications": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "List active medical qualifications",
                "description": "Returns all active medical qualifications sorted A→Z (e.g. MBBS, MD, BDS). Use qualification IDs to filter doctor listings. Requires HMAC authentication.",
                "operationId": "getQualifications",
                "responses": {
                    "200": {
                        "description": "Qualifications fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Qualifications fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Qualification"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        },
        "/api/v1/settings/public": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get public platform settings",
                "description": "Returns all settings flagged as public as a flat key → value object. Useful for configuring client apps (e.g. site name, support contact, feature flags). Requires HMAC authentication.",
                "operationId": "getPublicSettings",
                "responses": {
                    "200": {
                        "description": "Settings fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Settings fetched successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PublicSettings"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid HMAC headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HmacApiKey": [],
                        "HmacTimestamp": [],
                        "HmacNonce": [],
                        "HmacSignature": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "Standard error response body",
                "properties": {
                    "success": {
                        "description": "Reusable OpenAPI schema definitions.\nReferenced via $ref: '#/components/schemas/{name}'.",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message describing what went wrong"
                    }
                },
                "type": "object"
            },
            "Country": {
                "description": "ISO alpha-3 country",
                "properties": {
                    "code": {
                        "description": "ISO 3166-1 alpha-3 country code",
                        "type": "string",
                        "example": "IND",
                        "maxLength": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "India"
                    }
                },
                "type": "object"
            },
            "City": {
                "description": "City record",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mumbai"
                    },
                    "country_code": {
                        "type": "string",
                        "example": "IND"
                    }
                },
                "type": "object"
            },
            "Location": {
                "description": "Location / sub-city area",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "location": {
                        "type": "string",
                        "example": "Andheri West"
                    }
                },
                "type": "object"
            },
...
            "Service": {
                "description": "Parent service with nested child services",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "service": {
                        "type": "string",
                        "example": "Cardiology"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ServiceChild"
                        }
                    }
                },
                "type": "object"
            },
            "Qualification": {
                "description": "Medical qualification",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "qualification": {
                        "type": "string",
                        "example": "MBBS"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "HmacApiKey": {
                "type": "apiKey",
                "description": "Your client API key.",
                "name": "X-Api-Key",
                "in": "header"
            },
            "HmacTimestamp": {
                "type": "apiKey",
                "description": "Current Unix timestamp (UTC). Must be within ±5 minutes of server time.",
                "name": "X-Timestamp",
                "in": "header"
            },
            "HmacNonce": {
                "type": "apiKey",
                "description": "A random string (min 8 chars) unique per request.",
                "name": "X-Nonce",
                "in": "header"
            },
            "HmacSignature": {
                "type": "apiKey",
                "description": "HMAC-SHA256 hex signature. String to sign: \"METHOD\\nTIMESTAMP\\nNONCE\\nPATH\\nSHA256(body)\".",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}